Realtime/requirements2.txt
websocket-client
numpy
sounddevice
soundfile
openai
qdrant-client
librosa
wave
